% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_RCDscore.R
\name{cal_RCDscore}
\alias{cal_RCDscore}
\title{Calculate regulated cell death score using scRNA or bulk transcriptomic data}
\usage{
cal_RCDscore(
  expr,
  rcds = NULL,
  study.type = NULL,
  parallel.sz = NULL,
  verbose = FALSE
)
}
\arguments{
\item{expr}{log2 transformed TPM matrix derived from scRNAseq, or normalized gene expression matrix from bulk samples; cols are genes, rows are cells or samples.}

\item{rcds}{The list containing selected regulated cell death features.}

\item{study.type}{A string value to indicate the study type for calculation. Allowed values contain c('scRNAseq', 'bulk_RNAseq').}

\item{parallel.sz}{Number of processors to use when doing the calculations in parallel. This requires to previously load either the parallel or the snow library. If parallel is loaded and this argument is left with its default value (parallel.sz=0) then it will use all available core processors unless we set this argument with a smaller number. If snow is loaded then we must set this argument to a positive integer number that specifies the number of processors to employ in the parallel calculation.}

\item{verbose}{Gives information about each calculation step. Default: FALSE.}
}
\value{
Continuous numerical variables with RCDscore, named by cells or samples.
}
\description{
#' This is some description of this function.
}
\examples{
\dontrun{
data("exprset")
data("rcd.sig")
RCDscore <- cal_RCDscore(expr = t(exprset), rcds = rcd.sig, study.type = "bulk_RNAseq",parallel.sz = 12,verbose = T)
}
}
\author{
Wei Zhang
}
