% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModel.for.ICI.response.R
\name{CompareModel.for.ICI.response}
\alias{CompareModel.for.ICI.response}
\title{Using the 7 machine learning algorithms to construct a signature for ICI response}
\usage{
CompareModel.for.ICI.response(training, validation, method, sig)
}
\arguments{
\item{training}{Train model and tune parameters (10 times repeated, 5 folds cross-validation, 10 tuneLenth for each parameter)}

\item{validation}{Compare model performance and pick the best one as the final model}

\item{method}{A string value to indicate the study type for calculation. Allowed values contain c('scRNAseq', 'bulk_RNAseq').}

\item{sig}{Number of processors to use when doing the calculations in parallel. This requires to previously load either the parallel or the snow library. If parallel is loaded and this argument is left with its default value (parallel.sz=0) then it will use all available core processors unless we set this argument with a smaller number. If snow is loaded then we must set this argument to a positive integer number that specifies the number of processors to employ in the parallel calculation.}
}
\value{
Continuous numerical variables with RCDscore, named by cells or samples.
}
\description{
#' This is some description of this function.
}
