% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.OS.R
\name{prediction.OS}
\alias{prediction.OS}
\title{Calculation of the risk score of the overall survivla}
\usage{
prediction.OS(exp.matrix = NULL)
}
\arguments{
\item{exp.matrix}{A cohort matrix with the genes as the colnames, sample names as the rownamses. The colnames inlcudes the RCD.Sur.Sig, The top 3 colnames are "ID", "OS.time", "OS"}
}
\value{
A dataframe that contains the risk score
}
\description{
#' This is some description of this function.
}
\examples{
\dontrun{
data("fit")
data("rid")
data("exp_test")
dd = prediction.OS(exp.matrix = exp_test)

}
}
\author{
Wei Zhang
}
